Split out Image Compositions
